finally fixing nazis get vored text problem
the two ns didnt match and the letters were touching in nazis. this sucked ass, so i shrunk the heart at the end and fixed everything up good and made the world very happy.
also added a trans ant, a trans bee in milena's design, and an intersex bee also in milena's design.
also some snaps in this update
and deleting the greyspace WIP psd